Add developer tools for testing and rebasing.
bin/git can be used to run a git command on all the lab branches.
This makes rebasing on main significantly easier in the future
which will be helpful when upgrading versions. I didn't test the
case of a conflict, so have fun with that Future Tim.

bin/test can be used to run all unit tests, including the lab tests
for each branch. It's great for running the full test suite without
relying on CI.